Split ClientImportService
