+++ src/main/java/de/adorsys/keycloak/config/service/RealmImportService.java
@@ -74,6 +74,7 @@ public class RealmImportService {
     private final RequiredActionsImportService requiredActionsImportService;
     private final CustomImportService customImportService;
     private final ScopeMappingImportService scopeMappingImportService;
+    private final ClientAuthorizationImportService clientAuthorizationImportService;
     private final ClientScopeMappingImportService clientScopeMappingImportService;
     private final IdentityProviderImportService identityProviderImportService;
 
@@ -99,6 +100,7 @@ public class RealmImportService {
             RequiredActionsImportService requiredActionsImportService,
             CustomImportService customImportService,
             ScopeMappingImportService scopeMappingImportService,
+            ClientAuthorizationImportService clientAuthorizationImportService,
             ClientScopeMappingImportService clientScopeMappingImportService,
             IdentityProviderImportService identityProviderImportService,
             ChecksumService checksumService,
@@ -118,6 +120,7 @@ public class RealmImportService {
         this.requiredActionsImportService = requiredActionsImportService;
         this.customImportService = customImportService;
         this.scopeMappingImportService = scopeMappingImportService;
+        this.clientAuthorizationImportService = clientAuthorizationImportService;
         this.clientScopeMappingImportService = clientScopeMappingImportService;
         this.identityProviderImportService = identityProviderImportService;
         this.checksumService = checksumService;
@@ -194,6 +197,7 @@ public class RealmImportService {
         requiredActionsImportService.doImport(realmImport);
         authenticationFlowsImportService.doImport(realmImport);
         authenticatorConfigImportService.doImport(realmImport);
+        clientAuthorizationImportService.doImport(realmImport);
         clientImportService.doImportDependencies(realmImport);
         identityProviderImportService.doImport(realmImport);
         scopeMappingImportService.doImport(realmImport);
